@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import {MatDialog} from '@angular/material';
 
 @Component({
   selector: 'app-root',
@@ -8,5 +9,20 @@ import { Component } from '@angular/core';
 export class AppComponent {
   title = 'app';
 
-constructor(){}
+constructor(public dialog: MatDialog){}
+
+openDialog() {
+  const dialogRef = this.dialog.open(DialogContentExampleDialog);
+
+  dialogRef.afterClosed().subscribe(result => {
+    console.log(`Dialog result: ${result}`);
+  });
 }
+}
+
+@Component({
+  selector: 'dialog-content-example-dialog',
+  templateUrl: 'dialog-content-example-dialog.html',
+})
+export class DialogContentExampleDialog {}
+
